Refactor websocket server and client
- move single channel logic to webSocket struct
- split client and server into separate files
- add a utility optTicker
- add support for bidirectional ping pong
